+++ workers/unity/Packages/io.improbable.gdk.tools/.CodeGenTemplate/CodeGenerationLib/CodeWriter/Scopes/MethodBlock.cs
@@ -6,16 +6,16 @@ namespace Improbable.Gdk.CodeGeneration.CodeWriter.Scopes
 {
     public class MethodBlock : ScopeBody
     {
-        internal MethodBlock(string declaration, Action<MethodBlock> populate, string annotation = "") : base(declaration)
+        internal MethodBlock(string declaration, Action<MethodBlock> populate, IEnumerable<string> annotations = null) : base(declaration)
         {
-            Annotation = annotation;
+            Annotations = annotations;
             populate(this);
         }
 
-        internal MethodBlock(string declaration, Func<IEnumerable<string>> populate, string annotation = "") : base(
+        internal MethodBlock(string declaration, Func<IEnumerable<string>> populate, IEnumerable<string> annotations = null) : base(
             declaration)
         {
-            Annotation = annotation;
+            Annotations = annotations;
 
             var methodBody = populate().ToList();
             if (methodBody.Count > 0)